Moved helper methods
